requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
tqdm>=4.66.0